Separate running tests that run in the browser and those on the server
