+++ HISTORY.md
@@ -4,7 +4,7 @@ Unreleased
 
 0.8.10 / 2015-02-03
 ==================
-- Fixes documentation typos 
+- Fixes documentation typos
 - :green_heart: Fixes [#100](https://github.com/jhollingworth/marty/issues/100)
 - Fixes [#94](https://github.com/jhollingworth/marty/issues/94)
 - Fixes [#99](https://github.com/jhollingworth/marty/issues/99)
