+++ test/browser/addStoreChangeListenerSpec.js
@@ -6,7 +6,7 @@ describe('Marty#addStoreChangeListener()', function () {
   var Marty, Store1, Store2, listener, disposable;
 
   beforeEach(function () {
-    Marty = require('../index').createInstance();
+    Marty = require('../../index').createInstance();
     listener = sinon.spy();
 
     Store1 = Marty.createStore({
