+++ test/browser/lib/mockDispatcher.js
