+++ test/browser/storeFetchTest.js
@@ -1,6 +1,6 @@
 var sinon = require('sinon');
 var _ = require('underscore');
-var Marty = require('../index');
+var Marty = require('../../index');
 var expect = require('chai').expect;
 var Promise = require('es6-promise').Promise;
