+++ test/browser/storeSpec.js
@@ -1,12 +1,12 @@
 var sinon = require('sinon');
 var _ = require('lodash-node');
-var Marty = require('../index');
+var Marty = require('../../index');
 var expect = require('chai').expect;
-var Store = require('../lib/store');
-var Dispatcher = require('../dispatcher');
-var ActionPayload = require('../lib/actionPayload');
-var ActionHandlerNotFoundError = require('../errors/actionHandlerNotFound');
-var ActionPredicateUndefinedError = require('../errors/actionPredicateUndefined');
+var Store = require('../../lib/store');
+var Dispatcher = require('../../dispatcher');
+var ActionPayload = require('../../lib/actionPayload');
+var ActionHandlerNotFoundError = require('../../errors/actionHandlerNotFound');
+var ActionPredicateUndefinedError = require('../../errors/actionPredicateUndefined');
 
 describe('Store', function () {
   var store, changeListener, listener, dispatcher, dispatchToken = 'foo', initialState = {};
